Drop Scala 2.11
