+++ .github/workflows/ci.yml
@@ -40,7 +40,7 @@ jobs:
     strategy:
       fail-fast: false
       matrix:
-        scala: ['2.11.*', '2.12.*', '2.13.*', '3.*']
+        scala: ['2.12.*', '2.13.*', '3.*']
     steps:
       - name: Checkout current branch
         uses: actions/checkout@v3.0.2
@@ -58,9 +58,6 @@ jobs:
         uses: pierotofy/set-swap-space@master
         with:
           swap-size-gb: 7
-      - name: Test/compile 2.11
-        if: ${{ startsWith(matrix.scala, '2.11.') }}
-        run: free --si -tmws 10 & ./sbt ++${{ matrix.scala }} root211/Test/compile
       - name: Test/compile 2.12
         if: ${{ startsWith(matrix.scala, '2.12.') }}
         run: free --si -tmws 10 & ./sbt ++${{ matrix.scala }} root212/Test/compile
@@ -77,7 +74,7 @@ jobs:
     strategy:
       fail-fast: false
       matrix:
-        scala: ['2.11.*', '2.12.*', '2.13.*', '3.*']
+        scala: ['2.12.*', '2.13.*', '3.*']
     steps:
       - name: Checkout current branch
         uses: actions/checkout@v3.0.2
@@ -91,9 +88,6 @@ jobs:
         uses: coursier/cache-action@v6
       - name: Install libuv
         run: sudo apt-get update && sudo apt-get install -y libuv1-dev
-      - name: publishLocal 2.11
-        if: ${{ startsWith(matrix.scala, '2.11.') }}
-        run: ./sbt ++${{ matrix.scala }} root211/publishLocal
       - name: publishLocal 2.12
         if: ${{ startsWith(matrix.scala, '2.12.') }}
         run: ./sbt ++${{ matrix.scala }} root212/publishLocal
@@ -171,7 +165,7 @@ jobs:
     strategy:
       fail-fast: false
       matrix:
-        scala: ['2.11.12', '2.12.17', '2.13.10', '3.2.2']
+        scala: ['2.12.17', '2.13.10', '3.2.2']
         java: ['17']
         platform: ['JVM']
     steps:
@@ -186,11 +180,8 @@ jobs:
     - name: Cache scala dependencies
       uses: coursier/cache-action@v6
     - name: Mima Checks
-      if: ${{ !startsWith(matrix.scala, '3.') && !startsWith(matrix.scala, '2.11.') }}
+      if: ${{ !startsWith(matrix.scala, '3.') }}
       run: ./sbt ++${{ matrix.scala }}! mimaChecks
-    - name: Test 2.11
-      if: ${{ startsWith(matrix.scala, '2.11.') }}
-      run: ./sbt ++${{ matrix.scala }}! test${{ matrix.platform }}211
     - name: Test 2.12
       if: ${{ startsWith(matrix.scala, '2.12.') }}
       run: ./sbt ++${{ matrix.scala }}! test${{ matrix.platform }}
@@ -200,12 +191,6 @@ jobs:
     - name: Test 3
       if: ${{ startsWith(matrix.scala, '3.') }}
       run: ./sbt ++${{ matrix.scala }}! test${{ matrix.platform }}3
-    - name: Upload Test Results 2.11
-      if: ${{ startsWith(matrix.scala, '2.11.') }}
-      uses: actions/upload-artifact@v3
-      with:
-        name: zio-test-output-2.11
-        path: ./**/test-reports-zio/output.json
     - name: Upload Test Results 2.12
       if: ${{ startsWith(matrix.scala, '2.12.') }}
       uses: actions/upload-artifact@v3
@@ -218,12 +203,6 @@ jobs:
       with:
         name: zio-test-output-3
         path: ./**/test-reports-zio/output.json
-    - name: Report Test Death 2.11
-      if: ${{ failure() &&  startsWith(matrix.scala, '2.11.') }}
-      uses: actions/upload-artifact@v3
-      with:
-        name: zio-test-debug-2.11
-        path: ./**/test-reports-zio/**_debug.txt
     - name: Report Test Death 2.12
       if: ${{ failure() &&  startsWith(matrix.scala, '2.12.') }}
       uses: actions/upload-artifact@v3
