+++ build.sbt
@@ -4,6 +4,7 @@ import explicitdeps.ExplicitDepsPlugin.autoImport.moduleFilterRemoveValue
 import sbt.Keys
 
 Global / onChangedBuildSource := ReloadOnSourceChanges
+Global / concurrentRestrictions += Tags.limit(NativeTags.Link, 2)
 
 inThisBuild(
   List(
@@ -498,7 +499,6 @@ lazy val testMagnolia = crossProject(JVMPlatform, JSPlatform)
   .settings(crossProjectSettings)
   .settings(macroDefinitionSettings)
   .settings(
-    crossScalaVersions --= Seq(Scala211),
     scalacOptions ++= {
       if (scalaVersion.value == Scala3)
         Seq.empty
@@ -530,8 +530,7 @@ lazy val testMagnoliaTests = crossProject(JVMPlatform, JSPlatform)
   .dependsOn(testRunner)
   .settings(buildInfoSettings("zio.test"))
   .settings(
-    publish / skip := true,
-    crossScalaVersions --= Seq(Scala211)
+    publish / skip := true
   )
   .jsSettings(jsSettings)
   .enablePlugins(BuildInfoPlugin)
@@ -543,7 +542,6 @@ lazy val testRefined = crossProject(JVMPlatform, JSPlatform)
   .settings(crossProjectSettings)
   .settings(macroDefinitionSettings)
   .settings(
-    crossScalaVersions --= Seq(Scala211),
     libraryDependencies ++=
       Seq(
         ("eu.timepit" %% "refined" % "0.10.1").cross(CrossVersion.for3Use2_13)
@@ -557,7 +555,6 @@ lazy val testScalaCheck = crossProject(JSPlatform, JVMPlatform, NativePlatform)
   .settings(stdSettings("zio-test-scalacheck"))
   .settings(crossProjectSettings)
   .settings(
-    crossScalaVersions --= Seq(Scala211),
     libraryDependencies ++= Seq(
       ("org.scalacheck" %%% "scalacheck" % "1.16.0")
     )
@@ -631,7 +628,6 @@ lazy val testJunitRunnerTests = crossProject(JVMPlatform) // TODO: make plain pr
   .settings(publish / skip := true)
   .settings(testFrameworks += new TestFramework("zio.test.sbt.ZTestFramework"))
   .settings(
-    crossScalaVersions --= List(Scala211),
     libraryDependencies ++= Seq(
       "junit"                   % "junit"     % "4.13.2" % Test,
       "org.scala-lang.modules" %% "scala-xml" % "2.1.0"  % Test,
@@ -702,8 +698,7 @@ lazy val benchmarks = project.module
   .enablePlugins(JmhPlugin)
   .settings(replSettings)
   .settings(
-    // skip 2.11 benchmarks because akka stop supporting scala 2.11 in 2.6.x
-    crossScalaVersions --= List(Scala211, Scala3),
+    crossScalaVersions --= List(Scala3),
     //
     publish / skip := true,
     libraryDependencies ++=
@@ -753,7 +748,7 @@ val zioActorsV  = "0.0.9"
 lazy val scalafixSettings = List(
   scalaVersion := Scala213,
   addCompilerPlugin(scalafixSemanticdb),
-  crossScalaVersions --= List(Scala211, Scala212, Scala3),
+  crossScalaVersions --= List(Scala212, Scala3),
   scalacOptions ++= List(
     "-Yrangepos",
     "-P:semanticdb:synthetics:on"
@@ -819,7 +814,7 @@ lazy val docs = project.module
     Compile / fork := false,
     scalacOptions ~= { _ filterNot (_ startsWith "-Ywarn") },
     scalacOptions ~= { _ filterNot (_ startsWith "-Xlint") },
-    crossScalaVersions --= List(Scala211, Scala212, Scala3),
+    crossScalaVersions --= List(Scala212, Scala3),
     mdocIn  := (LocalRootProject / baseDirectory).value / "docs",
     mdocOut := (LocalRootProject / baseDirectory).value / "website" / "docs",
     ScalaUnidoc / unidoc / unidocProjectFilter := inProjects(
