+++ core-tests/shared/src/test/scala/zio/ExecutorSpec.scala
@@ -35,7 +35,7 @@ object TestExecutor {
     override def reportFailure(cause: Throwable): Unit = ()
   }
 
-  // backward compatibility for scala 2.11.12
+  // backward compatibility for scala 2.11.12 // TODO: remove?
   val runnable: Runnable = new Runnable {
     override def run(): Unit = ()
   }
