+++ project/plugins.sbt
@@ -9,8 +9,8 @@ addSbtPlugin("com.typesafe"                      % "sbt-mima-plugin"            
 addSbtPlugin("de.heikoseeberger"                 % "sbt-header"                    % "5.7.0")
 addSbtPlugin("org.portable-scala"                % "sbt-scala-native-crossproject" % "1.2.0")
 addSbtPlugin("org.portable-scala"                % "sbt-scalajs-crossproject"      % "1.2.0")
-addSbtPlugin("org.scala-js"                      % "sbt-scalajs"                   % "1.12.0")
-addSbtPlugin("org.scala-native"                  % "sbt-scala-native"              % "0.4.9")
+addSbtPlugin("org.scala-js"                      % "sbt-scalajs"                   % "1.13.0")
+addSbtPlugin("org.scala-native"                  % "sbt-scala-native"              % "0.4.10")
 addSbtPlugin("org.scalameta"                     % "sbt-mdoc"                      % "2.3.2")
 addSbtPlugin("org.scalameta"                     % "sbt-scalafmt"                  % "2.4.6")
 addSbtPlugin("pl.project13.scala"                % "sbt-jcstress"                  % "0.2.0")
