--- test/shared/src/main/scala-2.11/zio/test/PrettyPrintVersionSpecific.scala
@@ -1,10 +1,0 @@
-package zio.test
-
-import zio.stacktracer.TracingImplicits.disableAutoTrace
-
-trait PrettyPrintVersionSpecific {
-  def labels(product: Product): Iterator[String] = {
-    val _ = product
-    Iterator.continually("")
-  }
-}
